+++ HISTORY.rst
@@ -7,10 +7,12 @@ History
 - Add progress bar colors: blue while running, green on completion, red on failure (nens/rana-qgis-plugin#294)
 - Enforce .json suffix when saving importer presets to file (nens/threedi-schematisation-editor#493)
 - Validate settings files when importing via processing algorithm (nens/threedi-schematisation-editor#449)
+- Store draft settings upon closing importer wizard (nens/rana#4263)
 - Skip integration (with warning) when object matches to multiple conduits (nens/threedi-schematisation-editor#409)
 - Fix structures outside conduit raw bounding box not being matched despite being within snapping distance
 - Fix loading import settings; affects: point_to_line_conversion.lenght, point_to_line_conversion.azimuth, cross_section_location_mapping.join_field_src, join_field_src.join_field_tgt (nens/threedi-schematisation-editor#473)
 - Correct azimuth units in UI (was 'm', should be °)
+- Add reference_level field for cross section location import (nens/rana#3644)
 
 
 2.4.10 (2026-05-06)